Move code to package directory
