+++ entry_cli/calc_props.py
