+++ tests/context.py
@@ -2,4 +2,4 @@ import os
 import sys
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
-import calc_props
+from entry_cli import calc_props
